Code cleanup

+++ HoverTank1989/VehicleBase.cpp
@@ -69,7 +69,7 @@ void VehicleBase::RungeKutta4(struct VehicleData* aVehicle, double aTimeDelta)
     RightHandSide(aVehicle, &q, &dq1, aTimeDelta, 0.5, &dq2);
     RightHandSide(aVehicle, &q, &dq2, aTimeDelta, 0.5, &dq3);
     RightHandSide(aVehicle, &q, &dq3, aTimeDelta, 1.0, &dq4);
-    aVehicle->time = aVehicle->time + aTimeDelta;
+    aVehicle->time = aVehicle->time + static_cast<float>(aTimeDelta);
 
     DirectX::SimpleMath::Vector3 posUpdate = (dq1.position + 2.0 * dq2.position + 2.0 * dq3.position + dq4.position) / numEqns;
     DirectX::SimpleMath::Vector3 velocityUpdate = (dq1.velocity + 2.0 * dq2.velocity + 2.0 * dq3.velocity + dq4.velocity) / numEqns;
